#5 Created IndividualController and moved some functions to the new controller.
